requirements.txt
uvicorn
fastmcp
fastapi
pyjwt
pydantic
langchain-mcp-adapters
